+++ forth_src/io.fs
@@ -53,8 +53,8 @@ w ldx, lsb sta,x
 
 \ handle errors returned by open,
 \ close, and chkin. If ioresult is
-\ nonzero, close file and abort with
-\ an appropriate error message.
+\ nonzero, print error message and
+\ abort.
 : ioabort ( file# ioresult -- )
 ?dup if rvs case
 1 of ." too many files" endof
@@ -65,7 +65,6 @@ w ldx, lsb sta,x
 6 of ." not input file" endof
 7 of ." not output file" endof
 8 of ." missing filename" endof
-9 of ." illegal device number" endof
-." io err " dup .
-endcase clrchn close cr abort
-else drop then ;
+9 of ." illegal device #" endof
+." io err " endcase
+cr abort else drop then ;