+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
  - LOADB/SAVEB could change active device.
  - editor: Improved I/O error handling.
+ - IOABORT did not print all error messages.
 
 ## [3.0.0] - 2021-05-14